interface separada da main
