+++ Class/base.py
@@ -1,6 +1,7 @@
 import time
 
 from Class.shared import running, paused
+from Class.shared import clear_console
 
 class Base:
     def __init__(self, getter, threshold_percent, keys=None):
@@ -21,6 +22,8 @@ class Base:
                 except Exception as e:
                     print(f"Error in {self.__class__.__name__}: {e}. Restarting...")
                     continue
+                print(f"{self.getter}: {value} %")
+                #clear_console()
 
     def action(self):
         raise NotImplementedError
