+++ Class/shared.py
@@ -1,2 +1,6 @@
 running = True
 paused = False
+restart = False
+
+def clear_console():
+    print('\033c')
