+++ Json/coords.json
@@ -1,7 +1,7 @@
 {
     "coordinates": {
-        "coords_life": "1972, 1167, 2063, 1183",
-        "coords_pray": "2211, 1168, 2269, 1184",
-        "coords_pet_life": "3186, 957, 3323, 977"
+        "coords_life": "1524, 1168, 1615, 1184",
+        "coords_pray": "1763, 1169, 1821, 1185",
+        "coords_pet_life": "3186, 961, 3323, 981"
     }
 }
